made python3 version primary
